🧹 Jules05: Technical debt cleanup — architectural coherence
Scan and remediate architectural drift and code quality degradation:
- Relocated LSTMAttentionModel to src/models/lstm_model.py to consolidate sequence modeling logic.
- Unified Signal and ModelAction mappings using centralized ModelAction(idx).to_direction() across models, explainability, and benchmarks.
- Cleaned up raw print statements in main.py and research modules.
- Standardized model output layers to [HOLD, BUY, SELL] and removed legacy permutation logic.
- Established docs/technical-debt/DEBT_LOG.md for ongoing debt tracking.
