+++ docs/technical-debt/DEBT_LOG.md
@@ -1,45 +1,47 @@
 # Technical Debt Log
 
-This document tracks technical debt introduced by multi-agent parallelism and architectural drift.
+This log tracks architectural drift, code quality degradation, and fragmented logic introduced during multi-agent development.
 
-## Current Debt Items
+## Active Debt Items
 
-### Debt Item: Fragmented Model Weighting Logic
-**Category:** Fragmentation | Duplication
+### Debt Item: Fragmented LSTM Architecture
+**Category:** Fragmentation
 **Impact:** High
 **Effort:** M
-**Resolution plan:** Refactor `EnsembleModel` in `src/models/ensemble.py` to use `DynamicEnsemble` from `src/models/dynamic_ensemble.py`. `EnsembleModel` currently implements its own simpler (and slightly different) version of rolling Sharpe rebalancing.
-**Owner:** immediate cleanup (Jules05)
+**Resolution plan:** Relocate `LSTMAttentionModel` from `src/models/ensemble.py` to `src/models/lstm_model.py` to centralize sequence modeling logic.
+**Owner:** Jules05 (Immediate cleanup)
 
-### Debt Item: Inconsistent Signal Direction Mapping
-**Category:** Naming | Quality
-**Impact:** High
-**Effort:** S
-**Resolution plan:** Create a centralized `SignalDirection` or `TradeSignal` constants in `src/core/constants.py`.
-- `EnsembleModel`: `{0: 1, 1: -1, 2: 0}` (0=buy, 1=sell, 2=hold)
-- `PPOAgent` (via Adapter): `{0: 0, 1: 1, 2: -1}` (0=hold, 1=buy, 2=sell)
-- `TimeSeriesTransformer` (via Adapter): `{0: 1, 1: -1, 2: 0}` (0=buy, 1=sell, 2=hold)
-- `GymEnv`: `0=Hold, 1=Buy, 2=Sell`
-This inconsistency is a major source of potential bugs.
-**Owner:** immediate cleanup (Jules05)
-
-### Debt Item: Incomplete Type Hinting in PPOAgent
-**Category:** Quality
+### Debt Item: Redundant Signal/Action Mapping
+**Category:** Duplication
 **Impact:** Medium
 **Effort:** S
-**Resolution plan:** Add proper type hints to `src/models/ppo_agent.py` methods and `__init__`.
-**Owner:** immediate cleanup (Jules05)
+**Resolution plan:** Replace manual `IntEnum` to index mappings with centralized `ModelAction(idx).to_direction()` calls across the codebase.
+**Owner:** Jules05 (Immediate cleanup)
 
-### Debt Item: PR Residue and Debug Prints
-**Category:** Dead Code
+### Debt Item: Raw Print Statements in Research Modules
+**Category:** Quality
 **Impact:** Low
 **Effort:** S
-**Resolution plan:** Remove `print()` in `src/environment/gym_env.py` and replace with proper logging or `rich` console if needed.
-**Owner:** immediate cleanup (Jules05)
+**Resolution plan:** Replace `print()` with `structlog` or `rich.console` in `src/research/hyperopt_walkforward.py` and `main.py`.
+**Owner:** Jules05 (Immediate cleanup)
 
-### Debt Item: Missing Standard Model Interface
+### Debt Item: Placeholder RL Agents (Dreamer/PPO stubs)
+**Category:** Fragmentation
+**Impact:** Medium
+**Effort:** L
+**Resolution plan:** Schedule for Jules04 (Quant Research) to replace placeholders with functional world-model implementations.
+**Owner:** Jules04
+
+### Debt Item: Inconsistent Model Output Permutations
+**Category:** Fragmentation
+**Impact:** High
+**Effort:** M
+**Resolution plan:** Standardize all model outputs to `[HOLD, BUY, SELL]` and remove legacy permutation logic in `EnsembleModel`.
+**Owner:** Jules05 (Immediate cleanup)
+
+### Debt Item: Placeholder Reward Logic in Trading Environment
 **Category:** Quality
 **Impact:** Medium
 **Effort:** M
-**Resolution plan:** Define a `TradingModel` Protocol or Abstract Base Class to ensure all models (`PPOAgent`, `TimeSeriesTransformer`, `EnsembleModel`) share a common `predict` signature and metadata properties. This would eliminate the need for many adapters in `src/research/benchmarks.py`.
-**Owner:** Jules02 or Jules05 (Future)
+**Resolution plan:** Schedule for Jules01/Jules04 to implement proper reward shaping based on risk-adjusted returns.
+**Owner:** Jules01/Jules04
