+++ src/models/lstm_model.py
@@ -17,8 +17,57 @@ except ImportError:
     torch = None
     nn = None
 
-from src.core.constants import SignalDirection
+from src.core.constants import ModelAction, SignalDirection
 from src.models.base_model import BaseModel, Signal
+
+
+class LSTMAttentionModel(nn.Module if nn else object):
+    """
+    Bidirectional LSTM with multi-head self-attention.
+    Input : (batch, seq_len, n_features)
+    Output : (batch, 3) -> [hold_logit, buy_logit, sell_logit] (Standardized)
+    Note: Legacy checkpoints use [buy, sell, hold] and require permutation.
+    """
+
+    def __init__(
+        self,
+        n_features: int = 140,
+        hidden_size: int = 128,
+        num_layers: int = 2,
+        n_heads: int = 8,
+        dropout: float = 0.2,
+    ) -> None:
+        if not nn:
+            raise ImportError("PyTorch is required for LSTMAttentionModel")
+        super().__init__()
+        self.lstm = nn.LSTM(
+            input_size=n_features,
+            hidden_size=hidden_size,
+            num_layers=num_layers,
+            batch_first=True,
+            bidirectional=True,
+            dropout=dropout if num_layers > 1 else 0.0,
+        )
+        self.attn = nn.MultiheadAttention(
+            embed_dim=hidden_size * 2,
+            num_heads=n_heads,
+            dropout=dropout,
+            batch_first=True,
+        )
+        self.norm = nn.LayerNorm(hidden_size * 2)
+        self.head = nn.Sequential(
+            nn.Linear(hidden_size * 2, 64),
+            nn.GELU(),
+            nn.Dropout(dropout),
+            nn.Linear(64, 3),
+        )
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        out, _ = self.lstm(x)  # (B, T, 2*H)
+        attn_out, _ = self.attn(out, out, out)
+        out = self.norm(out + attn_out)  # residual
+        pooled = out.mean(dim=1)  # global average pool
+        return self.head(pooled)  # (B, 3)
 
 
 class LSTMPricePredictor(nn.Module if nn else object):
@@ -166,18 +215,11 @@ class LSTMModel(BaseModel):
                 logits = self.model(x)
                 probs = torch.softmax(logits, dim=-1).cpu().numpy()[0]
 
-            action_idx = np.argmax(probs)
+            action_idx = int(np.argmax(probs))
             confidence = float(probs[action_idx])
 
-            # Mapping: 0 -> HOLD, 1 -> BUY, 2 -> SELL
-            direction_map = {
-                0: SignalDirection.HOLD,
-                1: SignalDirection.BUY,
-                2: SignalDirection.SELL,
-            }
-
             return Signal(
-                direction=direction_map.get(action_idx, SignalDirection.HOLD),
+                direction=ModelAction(action_idx).to_direction(),
                 confidence=confidence,
                 metadata={
                     "probabilities": probs.tolist(),
